ui: extract settings dialog
